+++ defenses/__init__.py
@@ -9,6 +9,7 @@ from .tool_call_validator import ToolCallValidator
 from .cross_agent_boundary_guard import CrossAgentBoundaryGuard
 from .memory_isolation_guard import MemoryIsolationGuard
 from .tool_output_guard import ToolOutputGuard
+from .graph_boundary_guard import GraphBoundaryGuard
 from .prompt_hardening import LEVELS, get_prompt
 
 __all__ = [
@@ -31,4 +32,5 @@ __all__ = [
     "CrossAgentBoundaryGuard",
     "MemoryIsolationGuard",
     "ToolOutputGuard",
+    "GraphBoundaryGuard",
 ]